Moved ImageRenditionField to a more importable place
